+++ pkg/client/envconfig_unix.go
@@ -1,0 +1,8 @@
+//go:build !windows
+// +build !windows
+
+package client
+
+type OSSpecificEnv struct {
+	Shell string `env:"SHELL, parser=nonempty-string,default=/bin/bash"`
+}